apps/ml/requirements.txt
Flask==2.3.3
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
pytest==7.4.3
psycopg2-binary==2.9.9
flask-cors==4.0.0
python-dotenv==1.0.0
